requirements.txt
django>=3.2.*,<=4.0.*
django-picklefield~=3.0.1
croniter~=1.3.4
